Checkpoint realtime and timeline safety fixes

+++ Sources/Soundtime/AudioEditTimeline.swift
@@ -40,8 +40,8 @@ struct AudioEditTimeline: Sendable {
     }
 
     func frameRange(for selection: TimelineSelection) -> Range<Int> {
-        let startFrame = Int((selection.startProgress * Float(frameCount)).rounded(.down))
-        let endFrame = Int((selection.endProgress * Float(frameCount)).rounded(.up))
+        let startFrame = Int((selection.startProgress * Double(frameCount)).rounded(.down))
+        let endFrame = Int((selection.endProgress * Double(frameCount)).rounded(.up))
         return max(startFrame, 0)..<min(max(endFrame, startFrame), frameCount)
     }
